requirements.txt
allure-pytest==2.15.3
playwright==1.57.0
pytest==9.0.2